% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{DFM}
\alias{DFM}
\title{Determining the Number of Factors in the general dynamic factor model.}
\usage{
DFM(Y)
}
\arguments{
\item{Y}{p by n matrix of raw data, where p is the dimensionality, n is the sample size.}
}
\value{
estimated number of factors based on two information criterion in the General Dynamic Factor Model.
}
\description{
The function is used to minimize information criterion with two penalties in the general dynamic factor model. Comparing to the static factor model, parameters of information 
criterion in this model is not fixed, changing the fixed p,n and c to vectors.
}
\examples{
\dontrun{
p=100
n=100
Y<-array(rnorm(p*n),dim=c(p,n))
K<-DFM(Y)
}
}
