% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{FM}
\alias{FM}
\title{Determing the number of factors in approximate factor models.}
\usage{
FM(Y)
}
\arguments{
\item{Y}{p by n matrix of raw data, where p is the dimensionality, n is the sample size.}
}
\value{
estimated number of factors based on two infomation criterion using Bai & Ng method.
}
\description{
The function is adapted by Bai & Ng method to minimize information criterion with two penalties in static factor models.
}
\examples{
\dontrun{
p=100
n=100
Y<-array(rnorm(p*n),dim=c(p,n))
K<-FM(Y)
}
}
