% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_cpp}
\alias{gibbs_cpp}
\title{A Gibbs sampler using Rcpp}
\usage{
gibbs_cpp(N)
}
\arguments{
\item{N}{the number of samples}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
rnC <- gibbsC(100)
rnC
}
}
