\name{StatComp21031-package}
\alias{StatComp21031-package}
\alias{StatComp21031}
\docType{package}
\title{
  This package provides several functions using R and gives the functions to determine the number of factors in factor models.
}
\description{
    The \code{FM} function is adapted by Bai & Ng method to minimize information criterion with two penalties in static factor models.
    
    The \code{DFM} function is used to minimize information criterion with two penalties in the general dynamic factor model. Comparing to the static factor model, parameters of information criterion in this model is not fixed, changing the fixed p,n and c to vectors.
}
\author{
Siyu Wang

Maintainer: Siyu Wang <wsy0317@mail.ustc.edu.cn>
}

\keyword{ package }


