% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_cpp}
\alias{gibbs_cpp}
\title{A Gibbs sampler using Rcpp}
\usage{
gibbs_cpp(N)
}
\arguments{
\item{N}{the number of samples}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
rnC <- gibbsC(100,10)
par(mfrow=c(2,1));
plot(rnC[,1],type='l')
plot(rnC[,2],type='l')
}
}
